pyproject.toml
[project]
name = "kleine"
version = "0.1.0"
description = ""
authors = [
    {name = "Xavi"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"


[tool.poetry.dependencies]
python = "^3.11"
pyxavi = "^1.2.0"
python-dotenv = "^1.1"
spidev = [{version="^3.7", markers="sys_platform=='linux'"}]
smbus2 = "^0.5.0"
"rpi-lgpio" = [{version="^0.6", markers="sys_platform=='linux'"}]
gpiozero = [{version="^2.0", markers="sys_platform=='linux'"}]
pillow = "^11.2.1"
numpy = "^2.3.4"
lgpio = [{version="^0.2.2.0", markers="sys_platform=='linux'"}]
keyboard = "^0.13.5"

[tool.poetry.scripts]
main = "runner:run"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
